Store decimal answers and limits consistently as floats
round_number() returns int 0 for a zero input (rounding.py short-circuits
on `if num == 0`), and generate_csv.py stringified that result directly.
Any equation-based row whose answer or limit happened to be exactly zero
was therefore written as "0" while every other row of the same calculator
was written as a float. Six calculators ended up storing both forms.

Fixed generate_csv.py to coerce the answer and both limits to float
before stringifying (5 sites), and normalized the 168 affected fields
already in the data: 3 in test, 165 in train, 0 in one_shot, spanning
Delta Gap, Free Water Deficit, Delta Ratio, Anion Gap and Albumin
Corrected Delta Ratio.

round_number() itself is deliberately left alone: it is also used to
format numbers inside Ground Truth Explanation strings, so changing its
zero case would have rewritten explanation prose across the dataset. The
coercion is applied only where the answer and limit fields are built.

Scope is limited to Category in {lab test, physical, dosage}. Risk,
severity, diagnosis and date rows take the other branch of generate_csv
and are untouched -- integer risk scores stay integers.

Every write asserted numerically identical to the value it replaced.
Patient Note, Question, Relevant Entities, Ground Truth Explanation,
Output Type and Category are byte-identical across all three splits.
0 audit mismatches across all 11,693 rows, no calculator in these
categories still mixes int and float forms, and check_correctness grades
the normalized rows unchanged.

+++ calculator_implementations/generate_csv.py
@@ -128,7 +128,7 @@ for index, row in df_test.iterrows():
         print(row["Ground Truth Answer"], func_output["Answer"])
 
     if row["Category"] in ["lab test", "physical", "dosage"]:
-        csv_props["Ground Truth Answer"].append(str(round_number(float(func_output["Answer"]))))
+        csv_props["Ground Truth Answer"].append(str(float(round_number(float(func_output["Answer"])))))
     else:
         csv_props["Ground Truth Answer"].append(func_output["Answer"])
 
@@ -144,12 +144,12 @@ for index, row in df_test.iterrows():
             print("Lower Limit: ", round_number(answer_value * 1.05))
             print("Upper Limit: ", round_number(answer_value * 0.95))
 
-            csv_props["Lower Limit"].append(str(round_number(answer_value * 1.05)))
-            csv_props["Upper Limit"].append(str(round_number(answer_value * 0.95)))
+            csv_props["Lower Limit"].append(str(float(round_number(answer_value * 1.05))))
+            csv_props["Upper Limit"].append(str(float(round_number(answer_value * 0.95))))
 
         else:
-            csv_props["Lower Limit"].append(str(round_number(answer_value * 0.95)))
-            csv_props["Upper Limit"].append(str(round_number(answer_value * 1.05)))
+            csv_props["Lower Limit"].append(str(float(round_number(answer_value * 0.95))))
+            csv_props["Upper Limit"].append(str(float(round_number(answer_value * 1.05))))
